Split Helper Files into Core & Helper Files
The StenoByte_Core.c Source File is intended to remain consistent and compiled universally no matter the OS.
The StenoByte_Helper_for_Linux.c Source File is intended to be included when compiling for Linux, and excluded when compiling for other Operating Systems (OS's).

StenoByte_Core.h holds the definitions for StenoByte_Core.h and is the main dependency for the main.c file.
StenoByte_Helper.h holds the definitions for the Helper Source Files and will conditionally apply different declarations for different OS's.